+++ crates/fuel-core/src/database.rs
@@ -70,6 +70,8 @@ pub type Result<T> = core::result::Result<T, Error>;
 use crate::state::rocks_db::RocksDb;
 #[cfg(feature = "rocksdb")]
 use std::path::Path;
+use fuel_core_gas_price_service::fuel_gas_price_updater::fuel_core_storage_adapter::database::{ GasPriceMetadata};
+use crate::database::database_description::gas_price::GasPriceDatabase;
 
 // Storages implementation
 pub mod balances;
@@ -293,6 +295,16 @@ impl Modifiable for Database<OffChain> {
         commit_changes_with_height_update(self, changes, |iter| {
             iter.iter_all::<FuelBlockIdsToHeights>(Some(IterDirection::Reverse))
                 .map(|result| result.map(|(_, height)| height))
+                .try_collect()
+        })
+    }
+}
+
+impl Modifiable for Database<GasPriceDatabase> {
+    fn commit_changes(&mut self, changes: Changes) -> StorageResult<()> {
+        commit_changes_with_height_update(self, changes, |iter| {
+            iter.iter_all::<GasPriceMetadata>(Some(IterDirection::Reverse))
+                .map(|result| result.map(|(height, _)| height))
                 .try_collect()
         })
     }