+++ crates/fuel-core/src/service/adapters.rs
@@ -1,3 +1,18 @@
+use std::sync::Arc;
+
+use fuel_core_consensus_module::{
+    block_verifier::Verifier,
+    RelayerConsensusConfig,
+};
+use fuel_core_services::stream::BoxStream;
+#[cfg(feature = "p2p")]
+use fuel_core_types::services::p2p::peer_reputation::AppScore;
+use fuel_core_types::{
+    fuel_types::BlockHeight,
+    services::block_importer::SharedImportResult,
+};
+use fuel_core_upgradable_executor::executor::Executor;
+
 use crate::{
     database::{
         database_description::relayer::Relayer,
@@ -11,24 +26,12 @@ use crate::{
         vm_pool::MemoryPool,
     },
 };
-use fuel_core_consensus_module::{
-    block_verifier::Verifier,
-    RelayerConsensusConfig,
-};
-use fuel_core_services::stream::BoxStream;
-#[cfg(feature = "p2p")]
-use fuel_core_types::services::p2p::peer_reputation::AppScore;
-use fuel_core_types::{
-    fuel_types::BlockHeight,
-    services::block_importer::SharedImportResult,
-};
-use fuel_core_upgradable_executor::executor::Executor;
-use std::sync::Arc;
 
 pub mod block_importer;
 pub mod consensus_module;
 pub mod consensus_parameters_provider;
 pub mod executor;
+pub mod gas_price_adapters;
 pub mod graphql_api;
 #[cfg(feature = "p2p")]
 pub mod p2p;