+++ crates/services/gas_price_service/src/fuel_gas_price_updater/tests.rs
@@ -13,7 +13,7 @@ struct FakeL2BlockSource {
 
 #[async_trait::async_trait]
 impl L2BlockSource for FakeL2BlockSource {
-    async fn get_l2_block(&self, _height: BlockHeight) -> Result<BlockInfo> {
+    async fn get_l2_block(&mut self, _height: BlockHeight) -> Result<BlockInfo> {
         let block = self.l2_block.lock().await.recv().await.unwrap();
         Ok(block)
     }
@@ -23,7 +23,7 @@ struct PendingL2BlockSource;
 
 #[async_trait::async_trait]
 impl L2BlockSource for PendingL2BlockSource {
-    async fn get_l2_block(&self, _height: BlockHeight) -> Result<BlockInfo> {
+    async fn get_l2_block(&mut self, _height: BlockHeight) -> Result<BlockInfo> {
         futures::future::pending().await
     }
 }
@@ -42,64 +42,41 @@ impl FakeMetadata {
 
 #[async_trait::async_trait]
 impl MetadataStorage for FakeMetadata {
-    async fn get_metadata(&self) -> Result<Option<UpdaterMetadata>> {
+    async fn get_metadata(
+        &self,
+        _block_height: &BlockHeight,
+    ) -> Result<Option<UpdaterMetadata>> {
         Ok(self.inner.lock().await.clone())
     }
 
-    async fn set_metadata(&self, metadata: UpdaterMetadata) -> Result<()> {
+    async fn set_metadata(&mut self, metadata: UpdaterMetadata) -> Result<()> {
         let _ = self.inner.lock().await.replace(metadata);
         Ok(())
     }
 }
 
-fn arb_inner_updater() -> AlgorithmUpdaterV1 {
-    AlgorithmUpdaterV1 {
+fn arb_metadata() -> UpdaterMetadata {
+    UpdaterMetadata::V0(V0Metadata {
         // set values
         exec_gas_price_change_percent: 10,
         new_exec_price: 100,
         // unset values
         l2_block_height: 0,
         l2_block_fullness_threshold_percent: 0,
-        min_da_gas_price: 0,
-        max_da_gas_price_change_percent: 0,
-        total_da_rewards: 0,
-        da_recorded_block_height: 0,
-        latest_known_total_da_cost: 0,
-        projected_total_da_cost: 0,
-        da_p_component: 0,
-        da_d_component: 0,
-        profit_avg: 0,
-        avg_window: 0,
-        latest_da_cost_per_byte: 0,
-        last_da_gas_price: 0,
-        unrecorded_blocks: vec![],
         min_exec_gas_price: 0,
-    }
+    })
 }
 
-fn different_inner_updater() -> AlgorithmUpdaterV1 {
-    AlgorithmUpdaterV1 {
+fn different_arb_metadata() -> UpdaterMetadata {
+    UpdaterMetadata::V0(V0Metadata {
         // set values
         exec_gas_price_change_percent: 20,
         new_exec_price: 100,
         // unset values
         l2_block_height: 0,
         l2_block_fullness_threshold_percent: 0,
-        min_da_gas_price: 0,
-        max_da_gas_price_change_percent: 0,
-        total_da_rewards: 0,
-        da_recorded_block_height: 0,
-        latest_known_total_da_cost: 0,
-        projected_total_da_cost: 0,
-        da_p_component: 0,
-        da_d_component: 0,
-        profit_avg: 0,
-        avg_window: 0,
-        latest_da_cost_per_byte: 0,
-        last_da_gas_price: 0,
-        unrecorded_blocks: vec![],
         min_exec_gas_price: 0,
-    }
+    })
 }
 
 #[tokio::test]
@@ -108,8 +85,6 @@ async fn next__fetches_l2_block() {
     let l2_block = BlockInfo {
         height: 1,
         fullness: (60, 100),
-        block_bytes: 1000,
-        gas_price: 200,
     };
     let (l2_block_sender, l2_block_receiver) = tokio::sync::mpsc::channel(1);
     let l2_block_source = FakeL2BlockSource {
@@ -117,9 +92,9 @@ async fn next__fetches_l2_block() {
     };
     let metadata_storage = FakeMetadata::empty();
 
-    let inner = arb_inner_updater();
+    let starting_metadata = arb_metadata();
     let mut updater =
-        FuelGasPriceUpdater::init(inner.into(), l2_block_source, metadata_storage)
+        FuelGasPriceUpdater::init(starting_metadata, l2_block_source, metadata_storage)
             .await
             .unwrap();
 
@@ -131,32 +106,28 @@ async fn next__fetches_l2_block() {
     let new = next.await.unwrap().unwrap();
 
     // then
-    dbg!(&start, &new);
     assert_ne!(start, new);
 }
 
 #[tokio::test]
 async fn init__if_exists_already_reload() {
     // given
-    let metadata = arb_inner_updater();
-    let metadata_inner = Arc::new(Mutex::new(Some(metadata.into())));
+    let metadata = arb_metadata();
+    let metadata_inner = Arc::new(Mutex::new(Some(metadata.clone())));
     let metadata_storage = FakeMetadata {
         inner: metadata_inner,
     };
     let l2_block_source = PendingL2BlockSource;
 
     // when
-    let different_metadata = different_inner_updater();
-    let updater = FuelGasPriceUpdater::init(
-        different_metadata.into(),
-        l2_block_source,
-        metadata_storage,
-    )
-    .await
-    .unwrap();
+    let different_metadata = different_arb_metadata();
+    let updater =
+        FuelGasPriceUpdater::init(different_metadata, l2_block_source, metadata_storage)
+            .await
+            .unwrap();
 
     // then
-    let expected = arb_inner_updater();
+    let expected: AlgorithmUpdaterV0 = metadata.try_into().unwrap();
     let actual = updater.inner;
     assert_eq!(expected, actual);
 }
@@ -168,14 +139,14 @@ async fn init__if_it_does_not_exist_create_with_provided_values() {
     let l2_block_source = PendingL2BlockSource;
 
     // when
-    let metadata = different_inner_updater();
+    let metadata = different_arb_metadata();
     let updater =
-        FuelGasPriceUpdater::init(metadata.into(), l2_block_source, metadata_storage)
+        FuelGasPriceUpdater::init(metadata.clone(), l2_block_source, metadata_storage)
             .await
             .unwrap();
 
     // then
-    let expected = different_inner_updater();
+    let expected: AlgorithmUpdaterV0 = metadata.try_into().unwrap();
     let actual = updater.inner;
     assert_eq!(expected, actual);
 }
@@ -186,8 +157,6 @@ async fn next__new_l2_block_updates_metadata() {
     let l2_block = BlockInfo {
         height: 1,
         fullness: (60, 100),
-        block_bytes: 1000,
-        gas_price: 200,
     };
     let (l2_block_sender, l2_block_receiver) = tokio::sync::mpsc::channel(1);
     let l2_block_source = FakeL2BlockSource {
@@ -198,14 +167,11 @@ async fn next__new_l2_block_updates_metadata() {
         inner: metadata_inner.clone(),
     };
 
-    let mut inner = arb_inner_updater();
-    let mut updater = FuelGasPriceUpdater::init(
-        inner.clone().into(),
-        l2_block_source,
-        metadata_storage,
-    )
-    .await
-    .unwrap();
+    let inner = arb_metadata();
+    let mut updater =
+        FuelGasPriceUpdater::init(inner.clone(), l2_block_source, metadata_storage)
+            .await
+            .unwrap();
 
     // when
     let next = tokio::spawn(async move { updater.next().await });
@@ -214,15 +180,11 @@ async fn next__new_l2_block_updates_metadata() {
     let _ = next.await.unwrap().unwrap();
 
     // then
-    inner
-        .update_l2_block_data(
-            l2_block.height,
-            l2_block.fullness,
-            l2_block.block_bytes,
-            l2_block.gas_price,
-        )
+    let mut updater = AlgorithmUpdaterV0::try_from(inner).unwrap();
+    updater
+        .update_l2_block_data(l2_block.height, l2_block.fullness)
         .unwrap();
-    let expected = inner;
-    let actual = metadata_inner.lock().await.clone().unwrap().into();
+    let expected: UpdaterMetadata = updater.into();
+    let actual = metadata_inner.lock().await.clone().unwrap();
     assert_eq!(expected, actual);
 }