+++ crates/storage/src/structured_storage.rs
@@ -52,6 +52,7 @@ pub mod balances;
 pub mod blocks;
 pub mod coins;
 pub mod contracts;
+
 pub mod merkle_data;
 pub mod messages;
 pub mod sealed_block;